squash! test/Actors/Client/SubscriberTest.cs
Address the following finding.

#### ⚠️ Use documented derivation for "different number" — test/Actors/Client/SubscriberTest.cs#L70-L72

*Reported by: gemini, opus.*

fuzzy.instructions.md prescribes `x + fuzzy.SByte().Between(1, 5)` to derive a different numeric value. The test uses `fuzzy.Int32().Between(1, 5)`. Behaviorally equivalent, but the documented pattern is `SByte`.

```csharp
Assert.False(sut.Equals(new Subscriber(actorId, eventId + fuzzy.SByte().Between(1, 5), instance)));
```

+++ test/Actors/Client/SubscriberTest.cs
@@ -69,7 +69,7 @@ public abstract class SubscriberTest
 
         [Fact]
         public void ReturnsFalseWhenEventIdIsDifferent() =>
-            Assert.False(sut.Equals(new Subscriber(actorId, eventId + fuzzy.Int32().Between(1, 5), instance)));
+            Assert.False(sut.Equals(new Subscriber(actorId, eventId + fuzzy.SByte().Between(1, 5), instance)));
 
         [Fact]
         public void ReturnsFalseWhenInstanceIsDifferentReference() =>
